digital_bill/requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
idna==3.4
paystack==1.5.0
Pillow==9.5.0
pyotp==2.8.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
urllib3==2.0.7
